Add -f option to write each frame to separate files
- The -f option takes no arguments and signals that each frame should be
written to a separate file. If the output file name has a series of 0s
at the end before the file extension, they will be replaced by the frame
number. This will dynamically expand if the field width isn't large
enough for the frame number. If there are no 0s to replace, then the
frame number is inserted before the extension; if the input file has a
DPAN chunk, the frame count stored in it will be used to decide how many
0s to pad with. Note that 0s embedded in the middle of the filename will
not be replaced with the frame number. They must be at the end of the
name.
  Examples with the -f option:
    - For the output file name "funtime-00.gif", frame 1 will be written to "funtime-01.gif". Frame 12 will be written to "funtime-12.gif". Frame 1000 will be written to "funtime-1000.gif"
    - For the output file name "hello.gif" and the input file has a DPAN chunk that indicates 156 frames, frame 1 will be written to "hello01.gif". Frame 10 will be written to "hello10.gif". Frame 100 will be written to "hello100.gif".
    - For the output filename "world.gif" and the input file has no DPAN chunk, frame 1 will be written to "world1.gif". Frame 10 will be written to "world10.gif". Frame 100 will be written to "world100.gif". And so on.