Add the -c option to clip out only specific frames from the input.
The -c option clips out the specified frames from the input file and only writes
those to the output file. It accepts both a single frame or a range of frames of the form
*start*-*end*. If the start frame of a range is omitted, frame 1 is implied. If the end
frame of a range is omitted, the range continues to the final frame of the input file.
Multiple entries may be separated by commas.

Examples:
* -c -50
  Write the first 50 frames.
* -c 50-
  Skip the first 49 frames, and write all the rest.
* -c 10,20,30
  Only write frames 10, 20, and 30.
* -c 1,15-20
  Write frame 1 and frames 15-20.